requirements.txt
requests==2.31.0
pycryptodome==3.20.0
python-dotenv==1.0.0